feat: Enable users to specify the cache policy
