+++ src/raw/http_util/bytes_range.rs
@@ -19,6 +19,8 @@ use std::ops::Bound;
 use std::ops::RangeBounds;
 use std::str::FromStr;
 
+use bytes::Bytes;
+
 use crate::Error;
 use crate::ErrorKind;
 use crate::Result;
@@ -106,6 +108,27 @@ impl BytesRange {
                 None => Bound::Unbounded,
             },
         )
+    }
+
+    /// apply_on_bytes will apply range on bytes.
+    pub fn apply_on_bytes(&self, mut bs: Bytes) -> Bytes {
+        match (self.0, self.1) {
+            (None, None) => bs,
+            (None, Some(size)) => {
+                if size as usize >= bs.len() {
+                    return bs;
+                }
+                bs.split_off(bs.len() - size as usize)
+            }
+            (Some(offset), None) => bs.split_off(offset as usize),
+            (Some(offset), Some(size)) => {
+                let mut bs = bs.split_off(offset as usize);
+                if (size as usize) < bs.len() {
+                    let _ = bs.split_off(size as usize);
+                }
+                bs
+            }
+        }
     }
 }
 
@@ -267,6 +290,34 @@ mod tests {
             let actual = input.parse()?;
 
             assert_eq!(expected, actual, "{name}")
+        }
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_apply_on_bytes() -> Result<()> {
+        let bs = Bytes::from_static("Hello, World!".as_bytes());
+
+        let cases = vec![
+            ("full", (None, None), "Hello, World!"),
+            ("with_offset", (Some(1), None), "ello, World!"),
+            ("with_size", (None, Some(1)), "!"),
+            ("with_larger_size", (None, Some(100)), "Hello, World!"),
+            ("with_offset_and_size", (Some(1), Some(1)), "e"),
+            (
+                "with_offset_and_larger_size",
+                (Some(1), Some(100)),
+                "ello, World!",
+            ),
+            ("with_empty_offset", (Some(0), Some(100)), "Hello, World!"),
+        ];
+
+        for (name, input, expected) in cases {
+            let actual = BytesRange(input.0, input.1).apply_on_bytes(bs.clone());
+            let actual = String::from_utf8_lossy(&actual);
+
+            assert_eq!(expected, &actual, "{name}");
         }
 
         Ok(())
